make-cards: Remove the points scale
Pango seems to get confused about the kerning when the transformation
matrix is too complicated. If we remove the scaling then the matrix
should only have rotations and translations. Previously there was an
attempt to counteract the scale just before rendering the Pango layouts
but maybe this wasn’t working because when combined with the rotation it
was no longer able to recognise that the matrix is simple.